Fix Linter and doc
