+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Awesome Books</title>
     <link rel="stylesheet" href="styles.css">
-    <script src="src/index.js" defer></script>
+    <script type="module"src="src/index.js" defer></script>
 </head>
 <body>
     <section class="area">
