+++ styles.css
@@ -1,10 +1,13 @@
 body {
-  margin: 0px auto;
+  margin: 0 auto;
 }
 
 ul {
   list-style-type: none;
   padding: 8px;
+  display: none;
+  width: 300px;
+  border: 2px solid black;
 }
 
 .area {
@@ -15,27 +18,20 @@ ul {
   width: 100%;
 }
 
-ul {
-  width: 300px;
-  border: 2px solid black;
-}
-
 li {
-  
-height: 45px;
+  height: 45px;
 }
 
 button {
   width: 70px;
-height: 30px;
+  height: 30px;
 }
 
-li p{
+li p {
   float: left;
   margin-top: 6px;
 }
 
-li button{
+li button {
   float: right;
 }
-
